+++ programs/drift/src/controller/token.rs
@@ -1,6 +1,6 @@
 use crate::signer::get_signer_seeds;
 use anchor_lang::prelude::*;
-use anchor_spl::token::{self, Token, TokenAccount, Transfer};
+use anchor_spl::token::{self, CloseAccount, Token, TokenAccount, Transfer};
 
 pub fn send_from_program_vault<'info>(
     token_program: &Program<'info, Token>,
@@ -37,4 +37,23 @@ pub fn receive<'info>(
     let cpi_program = token_program.to_account_info();
     let cpi_context = CpiContext::new(cpi_program, cpi_accounts);
     token::transfer(cpi_context, amount)
+}
+
+pub fn close_vault<'info>(
+    token_program: &Program<'info, Token>,
+    account: &Account<'info, TokenAccount>,
+    destination: &AccountInfo<'info>,
+    authority: &AccountInfo<'info>,
+    nonce: u8,
+) -> Result<()> {
+    let signature_seeds = get_signer_seeds(&nonce);
+    let signers = &[&signature_seeds[..]];
+    let cpi_accounts = CloseAccount {
+        account: account.to_account_info().clone(),
+        destination: destination.clone(),
+        authority: authority.to_account_info().clone(),
+    };
+    let cpi_program = token_program.to_account_info();
+    let cpi_context = CpiContext::new_with_signer(cpi_program, cpi_accounts, signers);
+    token::close_account(cpi_context)
 }