+++ programs/drift/src/instructions/keeper.rs
@@ -946,7 +946,7 @@ pub fn handle_resolve_perp_bankruptcy(
     )?;
 
     validate!(
-        quote_spot_market_index == 0,
+        quote_spot_market_index == QUOTE_SPOT_MARKET_INDEX,
         ErrorCode::InvalidSpotMarketAccount
     )?;
 
@@ -1168,13 +1168,6 @@ pub fn handle_update_funding_rate(
         "Market funding is paused",
     )?;
 
-    validate!(
-        ((clock_slot == perp_market.amm.last_update_slot && perp_market.amm.last_oracle_valid)
-            || perp_market.amm.curve_update_intensity == 0),
-        ErrorCode::AMMNotUpdatedInSameSlot,
-        "AMM must be updated in a prior instruction within same slot"
-    )?;
-
     let funding_paused =
         state.funding_paused()? || perp_market.is_operation_paused(PerpOperation::UpdateFunding);
 
@@ -1298,6 +1291,19 @@ pub fn handle_update_perp_bid_ask_twap(ctx: Context<UpdatePerpBidAskTwap>) -> Re
         perp_market.amm.last_ask_price_twap,
         perp_market.amm.last_mark_price_twap_ts
     );
+
+    let funding_paused =
+        state.funding_paused()? || perp_market.is_operation_paused(PerpOperation::UpdateFunding);
+    controller::funding::update_funding_rate(
+        perp_market.market_index,
+        perp_market,
+        &mut oracle_map,
+        now,
+        slot,
+        &state.oracle_guard_rails,
+        funding_paused,
+        None,
+    )?;
 
     Ok(())
 }