+++ sdk/src/types.ts
@@ -999,10 +999,19 @@ export type ReferrerInfo = {
 	referrerStats: PublicKey;
 };
 
-export type TxParams = {
+export type BaseTxParams = {
 	computeUnits?: number;
 	computeUnitsPrice?: number;
 };
+
+export type ProcessingTxParams = {
+	useSimulatedComputeUnits?: boolean;
+	computeUnitsBufferMultiplier?: number;
+	useSimulatedComputeUnitsForCUPriceCalculation?: boolean;
+	getCUPriceFromComputeUnits?: (computeUnits: number) => number;
+};
+
+export type TxParams = BaseTxParams & ProcessingTxParams;
 
 export class SwapReduceOnly {
 	static readonly In = { in: {} };